requirements.txt
assertpy==1.1
paho-mqtt==1.5.1
prometheus-client==0.10.0
pyserial==3.5
pytest==6.2.2
toml==0.10.2
transitions==0.8.7
